Move UnlockConfirmation & LockedByButton to a new file
