+++ src/features/CourtCaseList/tags/LockedByTag/LockedByTag.tsx
@@ -1,105 +1,7 @@
 import ConditionalRender from "components/ConditionalRender"
-import Image from "next/image"
 import { useState } from "react"
-import { createUseStyles } from "react-jss"
-import { useCustomStyles } from "../../../../../styles/customStyles"
+import LockedByButton from "./LockedByButton"
 import LockedByText from "./LockedByText"
-
-const useStyles = createUseStyles({
-  LockedIcon: {
-    // Change colour from black to GDS text-blue (#144e81) with CSS filters
-    filter: "invert(12%) sepia(70%) saturate(4629%) hue-rotate(197deg) brightness(97%) contrast(84%)"
-  },
-  LockedByURL: {
-    textDecoration: "underline",
-    marginTop: 4,
-    marginBottom: 2,
-    fontWeight: "normal",
-    border: "none",
-    outline: "none",
-    background: "none",
-    cursor: "pointer",
-    fontFamily: "inherit",
-    fontSize: "inherit",
-    color: "inherit",
-    letterSpacing: "0.5px",
-    textTransform: "none"
-  }
-})
-
-interface UnlockConfirmationProps {
-  onCancel: () => void
-  unlockPath?: string
-}
-
-const UnlockConfirmation = ({ onCancel, unlockPath }: UnlockConfirmationProps) => {
-  return (
-    <>
-      <p>{"Click the button to unlock the case"}</p>
-      <div className="govuk-button-group">
-        <form method="POST" action={unlockPath}>
-          <button className="govuk-button" data-module="govuk-button" id="unlock">
-            {"Unlock"}
-          </button>
-        </form>
-        <a
-          className="govuk-link"
-          href="/"
-          onClick={(event) => {
-            event.preventDefault()
-            onCancel()
-          }}
-        >
-          {"Cancel"}
-        </a>
-      </div>
-    </>
-  )
-}
-
-interface LockedByButtonProps {
-  lockedBy?: string | null
-  unlockPath?: string
-  showUnlockConfirmation: boolean
-  setShowUnlockConfirmation: (value: boolean) => void
-}
-
-const LockedByButton = ({
-  lockedBy,
-  unlockPath,
-  showUnlockConfirmation,
-  setShowUnlockConfirmation
-}: LockedByButtonProps) => {
-  const classes = useStyles()
-  const lockedByButtonClasses = useCustomStyles()
-  return (
-    <>
-      <button
-        className={lockedByButtonClasses["button--tag"]}
-        onClick={() => {
-          setShowUnlockConfirmation(true)
-        }}
-      >
-        <Image
-          src={"/bichard/assets/images/lock.svg"}
-          width={18}
-          height={18}
-          className={unlockPath ? classes.LockedIcon : undefined}
-          alt="Lock icon"
-        />
-        {lockedBy}
-      </button>
-      {showUnlockConfirmation && (
-        <UnlockConfirmation
-          onCancel={() => {
-            setShowUnlockConfirmation(false)
-          }}
-          unlockPath={unlockPath}
-        />
-      )}
-    </>
-  )
-}
 
 interface LockedByTagProps {
   lockedBy?: string | null
